# [Enhancement] Disable operator's mem_tracker if profile is disable