Signed-off-by: liuyehcf <1559500551@qq.com>

## What type of PR is this：
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->

#17023 bring performance deduction, the reason comes from two aspects:
1. 3 atomic rather than 1 atomic are updated each time for MemTracker updation.
2. there have been more context switch. when calling `push_chunk` and `pull_chunk`, the thread_local will change to operator's mem_tracker, in memory intensive cases, the total amount of context switch will be tremendous, which will finally lead to MemTrack's updation.

## Experiment

benchmark_04(1fe, 3be, 16c/64G)

```sql
select count(*) from ssb_200g.lineorder_flat where LO_ORDERDATE != '1994-04-12';
```

|  | 1 concurrency | 10 concurrency |
|:--|:--|:--|
| before #17023(disable profile) | 177ms | 0.716s |
| before #17023(enable profile) | 183ms | 0.7.66s |
| #17023(disable profile) | 417ms | 1.596s |
| #17023(enable profile) | 435ms | 1.596s |
| this pr(disable profile) | 176ms | 0.699s |
| this pr(enable profile) | 390ms | 1.317s |

So we can draw the conclusion that this pr's performance is same as before #17023 if profile is disabled. But there still has substantial deduction in performance if profile is enabled for memory intensive situations.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be auto backported to target branch
  - [ ] 2.5
  - [ ] 2.4
  - [ ] 2.3
  - [ ] 2.2