@wanpengfei-git 2023-02-11
#### [FE PR Coverage Check]

:disappointed: **fail** : 0 / 7 (00.00%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/qe/Coordinator.java|0|7|00.00%|[1626, 1627, 1665, 1666, 1667, 1668, 1669]|