@wanpengfei-git 2023-02-13
run starrocks_admit_test